[DMS-1124] Slice 3 task 1: Extract shared classification and key-unification cores
Extracts per-table classification and key-unification logic from
ProfileRootTableBindingClassifier and ProfileRootKeyUnificationResolver
into internal cores so slice 3 separate-table classifier can share the
same invariant logic. Renames BuildCurrentRootRowByColumnName to
BuildCurrentRowByColumnName since the method is already tableModel-generic.
Generalizes BuildCandidateScopeSet to ImmutableArray<string> so the
longest-first ordering invariant is type-enforced and reused without
re-materialization. Reworded core exception messages to drop the
root-table prefix (the cores are scope-family-agnostic).

No behavior change; slice 2 unit + integration tests pass with only
three test-assertion wording updates to match the reworded exceptions.
